+++ src/session.rs
@@ -2,8 +2,8 @@ use chrono::{DateTime, Duration, Utc};
 use rand::Rng;
 
 pub struct Session {
-    pub user_id: u64,
-    pub session_id: u64,
+    pub user_id: u32,
+    pub session_id: u32,
     pub created_at: DateTime<Utc>,
     pub expires_at: DateTime<Utc>,
 }
@@ -11,9 +11,9 @@ pub struct Session {
 const SESSION_LIFETIME: chrono::TimeDelta = Duration::hours(1);
 
 impl Session {
-    pub fn new(user_id: u64) -> Self {
+    pub fn new(user_id: u32) -> Self {
         let mut random_number = rand::thread_rng();
-        let session_id: u64 = random_number.gen();
+        let session_id: u32 = random_number.gen();
         let now = Utc::now();
         Session {
             user_id,
